Write jest tests
