+++ apps/server/src/schema/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./authTypes";
+export * from "./authMutation";
