+++ apps/server/src/schema/index.ts
@@ -1,4 +1,4 @@
 export * from "./photo";
 export * from "./album";
 export * from "./query";
-export * from "./authentication";
+export * from "./auth";
