+++ apps/webapp/codegen.yml
@@ -1,6 +1,6 @@
 overwrite: true
 schema: "./node_modules/@howdypix/graphql-schema/schema.json"
-documents: ./pages/**/*.tsx
+documents: ./src/**/*.tsx
 generates:
   src/__generated__/schema-types.ts:
     plugins:
